requirements.txt
vonage
websocket-client
numpy
scipy
fastapi
uvicorn[standard]